requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
tenacity>=8.0.0
rich>=13.0.0
typer>=0.12.0
aiofiles>=23.0.0
requests>=2.31.0
